requirements.txt
ultralytics==8.3.236
torch
opencv-python
numpy
Pillow
tkinterdnd2
